+++ resources/views/admin/dashboard/item.blade.php
@@ -1,5 +1,9 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
 <!-- dashboard item -->
 =======
 notify
 >>>>>>> 946fdba366 (Squashed 'laravel/Modules/Notify/' content from commit 6aac1e028)
+=======
+<!-- dashboard item -->
+>>>>>>> 1c8d7d06e0 (Squashed 'laravel/Modules/Tenant/' content from commit be731f696)