+++ vite.config.js
@@ -1,9 +1,15 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
 const dotenvExpand = require('dotenv-expand');
 dotenvExpand(require('dotenv').config({ path: '../../.env'/*, debug: true*/ }));
 
 =======
 >>>>>>> 38c1507055 (Squashed 'laravel/Modules/Media/' content from commit 4548be09a)
+=======
+const dotenvExpand = require('dotenv-expand');
+dotenvExpand(require('dotenv').config({ path: '../../.env'/*, debug: true*/ }));
+
+>>>>>>> 1c8d7d06e0 (Squashed 'laravel/Modules/Tenant/' content from commit be731f696)
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
 
@@ -11,10 +17,14 @@ export default defineConfig(
     {
         build: {
 <<<<<<< HEAD
+<<<<<<< HEAD
             outDir: '../../public/build-user',
 =======
             outDir: '../../public/build-media',
 >>>>>>> 38c1507055 (Squashed 'laravel/Modules/Media/' content from commit 4548be09a)
+=======
+            outDir: '../../public/build-user',
+>>>>>>> 1c8d7d06e0 (Squashed 'laravel/Modules/Tenant/' content from commit be731f696)
             emptyOutDir: true,
             manifest: true,
         },
@@ -23,10 +33,14 @@ export default defineConfig(
                 {
                     publicDirectory: '../../public',
 <<<<<<< HEAD
+<<<<<<< HEAD
                     buildDirectory: 'build-user',
 =======
                     buildDirectory: 'build-media',
 >>>>>>> 38c1507055 (Squashed 'laravel/Modules/Media/' content from commit 4548be09a)
+=======
+                    buildDirectory: 'build-user',
+>>>>>>> 1c8d7d06e0 (Squashed 'laravel/Modules/Tenant/' content from commit be731f696)
                     input: [
                         __dirname + '/resources/assets/sass/app.scss',
                         __dirname + '/resources/assets/js/app.js'
@@ -38,6 +52,7 @@ export default defineConfig(
     }
 );
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 
 //export const paths = [
@@ -45,3 +60,5 @@ export default defineConfig(
 //    'Modules/$STUDLY_NAME$/resources/assets/js/app.js',
 //];
 >>>>>>> 38c1507055 (Squashed 'laravel/Modules/Media/' content from commit 4548be09a)
+=======
+>>>>>>> 1c8d7d06e0 (Squashed 'laravel/Modules/Tenant/' content from commit be731f696)